Update dwarf.cpp
This is how roles should be calculated, excluding aspects with no elements.

It's a horrible hack be checking for 50.0f vs going back and removing the value beforehand (maybe another value should be used to flag it?).  Because in the small chance a value is actually 50% (example if a role hits medians on all attributes and it's attribute only).